# openapi.yaml
openapi: 3.1.0
info:
  title: Koli Parts API
  version: 0.1.0
  description: >-
    Contract-first MVP API. Supplier procurement endpoints are internal abstractions;
    they do not imply eBay Order API production access.
servers:
  - url: http://localhost:4000
    description: local
  - url: https://api.koli-one.com
    description: production-proposed
security:
  - cookieAuth: []
...
paths:
...
  /auth/koli-one/exchange:
    post:
      tags: [Auth]
      security: []
      summary: Exchange a verified short-lived Koli One identity assertion for a Koli Parts session
      operationId: exchangeKoliOneAssertion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assertion]
              properties:
                assertion: { type: string, minLength: 20 }
      responses:
        '200':
          description: Session established
          headers:
            Set-Cookie:
              description: Secure HttpOnly Koli Parts session cookie
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Return the current Koli Parts session user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current session user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: kp_session
...
  responses:
    Unauthorized: { description: Unauthorized }
...
  schemas:
    AuthSession:
      type: object
      required: [userId, roles]
      properties:
        userId: { type: string, format: uuid }
        preferredLanguage: { type: string, enum: [bg, en] }
        roles:
          type: array
          items: { type: string }
        expiresAt: { type: string, format: date-time }
...